validator bugs fixed
