+++ validators/category-validator.js
@@ -6,16 +6,4 @@ exports.generateValidator = [
     .trim()
     .isLength({ min: 1 })
     .escape(),
-  (req, res, next) => {
-    const errors = validationResult(req);
-    if (!errors.isEmpty()) {
-      res.render("category-form", {
-        title: "Create category",
-        category: req.body,
-        errors: errors.array(),
-      });
-      return;
-    }
-    next();
-  },
 ];
