+++ validators/item-validator.js
@@ -3,6 +3,7 @@ const Category = require("../models/category");
 
 exports.generateValidator = [
   body("name", "Name must not be empty.").trim().isLength({ min: 1 }).escape(),
+  body("category", "Select a category").isLength({ min: 1 }),
   body("description", "Description must not be empty.")
     .trim()
     .isLength({ min: 1 })
@@ -15,21 +16,4 @@ exports.generateValidator = [
     .trim()
     .optional({ checkFalsy: true })
     .isNumeric(),
-  (req, res, next) => {
-    const errors = validationResult(req);
-    if (!errors.isEmpty()) {
-      Category.find({}).exec(function (err, results) {
-        if (err) return next(err);
-
-        res.render("item-form", {
-          title: "Create item",
-          item: req.body,
-          categories: results,
-          errors: errors.array(),
-        });
-      });
-      return;
-    }
-    next();
-  },
 ];
